moved test to modules
