+++ Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-anyhow = "1.0.98"
 chrono = { version = "0.4.41", features = ["serde"] }
 serde = { version = "1.0.219", features = ["derive"] }
 diesel = { version = "2.2.11", features = ["sqlite", "r2d2", "chrono", "returning_clauses_for_sqlite_3_35"] }
